Extract ConfigModule. Add console log appender for warnings and above. Always return same instances from ConfigStore.

+++ Sonneville.Utilities/Configuration/NiniConfigStore.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Nini.Config;
@@ -17,20 +18,27 @@ namespace Sonneville.Utilities.Configuration
     public class NiniConfigStore : INiniConfigStore
     {
         private readonly string _path;
+        private readonly Dictionary<Type, object> _configs;
 
         public NiniConfigStore(string path)
         {
             _path = path;
+            _configs = new Dictionary<Type, object>();
         }
 
         public void Save<T>(T config)
         {
+            var type = typeof(T);
+            if (_configs.TryGetValue(type, out var existing) && !ReferenceEquals(existing, config))
+            {
+                throw new ArgumentOutOfRangeException(nameof(config), config, "Must pass original config instance!");
+            }
             if (!File.Exists(_path)) File.Create(_path).Close();
-            var configSource = new IniConfigSource(_path);
 
+            var configSource = new IniConfigSource(_path);
             var section = configSource.Configs[GetSectionForType<T>()] ?? configSource.AddConfig(GetSectionForType<T>());
 
-            typeof(T).GetProperties()
+            type.GetProperties()
                 .ToDictionary(info => info.Name, info => info.GetValue(config))
                 .Where(kvp => kvp.Value != null).ToList()
                 .ForEach(kvp => section.Set(kvp.Key, kvp.Value));
@@ -39,11 +47,17 @@ namespace Sonneville.Utilities.Configuration
 
         public T Read<T>() where T : new()
         {
-            var config = new T();
+            var type = typeof(T);
+            if (!_configs.ContainsKey(type))
+            {
+                _configs.Add(type, new T());
+            }
+
+            var config = (T) _configs[type];
             if (File.Exists(_path))
             {
                 var configSource = new IniConfigSource(_path);
-                foreach (var propertyInfo in typeof(T).GetProperties())
+                foreach (var propertyInfo in type.GetProperties())
                 {
                     var stringValue = configSource.Configs[GetSectionForType<T>()].Get(propertyInfo.Name);
                     dynamic value = Convert.ChangeType(stringValue, propertyInfo.PropertyType);
